refactor(core): rename events folder to event
