+++ src/domain/file/event/file-removed/actionCreators.ts
