+++ src/domain/file/event/file-removed/fileRemoved.ts
