+++ src/domain/file/event/file-stored/fileStored.ts
